RAG_PIPELINE/requirements.txt
fastapi
uvicorn
chromadb
langchain
langchain-community
langchain-google-genai
pypdf
python-multipart
pydantic-settings
uvloop
httptools
presidio-analyzer
presidio-anonymizer
spacy
langgraph
tavily-python
